Breaking up DecisionList component. Decision component was born.
