+++ dashboard-ui/src/components/ActionsCompare/decisionList.jsx
@@ -1,22 +1,10 @@
 import React from 'react';
-import { ListGroup } from 'react-bootstrap';
 import Accordion from 'react-bootstrap/Accordion';
-import { getCasualtyArray } from './htmlUtility';
 import { Query } from 'react-apollo';
 import gql from 'graphql-tag';
+import * as utility from './utility'
+import Decision from './decision';
 
-const getScenarioNamesQueryName = "getScenarioNames";
-const getPerformerADMByScenarioName = "getPerformerADMsForScenario";
-const getTestByADMandScenarioName = "getTestByADMandScenario";
-
-const scenario_names_aggregation = gql`
-    query getScenarioNames{
-        getScenarioNames
-    }`;
-const performer_adm_by_scenario = gql`
-    query getPerformerADMsForScenario($scenarioID: ID){
-        getPerformerADMsForScenario(scenarioID: $scenarioID)
-    }`;
 const test_by_adm_and_scenario = gql`
     query getTestByADMandScenario($scenarioID: ID, $admName: ID){
         getTestByADMandScenario(scenarioID: $scenarioID, admName: $admName)
@@ -37,133 +25,6 @@ class DecisionList extends React.Component {
     };
   }
 
-  formattedActionType(actionType) {
-    return actionType
-      .split('_')
-      .map(word => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-      .join(' ');
-  }
-
-  tagMappings = {
-    "red": "Immediate",
-    "gray": "Expectant",
-    "green": "Minimal",
-    "yellow": "Delayed"
-  }
-
-  admCommandMap = {
-    "Tag Casualty": "Tag Applied",
-    "Apply Treatment": "Injury Treated",
-    "Check All Vitals": "Check Vitals",
-    "Move to EVAC": "Move to Evac"
-  }
-
-  humanActionMap = {
-    "Pulse Taken": "Check Vitals",
-    "Tag Applied": "Tag Applied",
-    "Injury Treated": "Injury Treated",
-    "Move To Evac": "Move To Evac"
-  }
-
-  admPulseMapping(hrpmin) {
-    if (hrpmin > 0 && hrpmin <= 50) {
-      return "Faint Pulse";
-    } else if (hrpmin > 50 && hrpmin <= 100) {
-      return "Normal Pulse";
-    } else if (hrpmin > 100) {
-      return "Fast Pulse";
-    } else {
-      return "No Pulse";
-    }
-  }
-
-
-  humanPulseReading(pulse) {
-    switch (pulse) {
-      case "pulse_fast":
-        return "Fast Pulse"
-      case "pulse_normal":
-        return "Normal Pulse"
-      case "pulse_faint":
-        return "Faint Pulse"
-      default:
-        return "unknown"
-    }
-  }
-
-  renderDecisionCSV = (decision) => {
-    switch (decision.actionType) {
-      case "PULSE_TAKEN":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {decision.actionData[1]}</ListGroup.Item>
-            <ListGroup.Item>Pulse Reading: {this.humanPulseReading(decision.actionData[0])}</ListGroup.Item>
-          </div>
-        );
-      case "INJURY_TREATED":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {decision.actionData[1]}</ListGroup.Item>
-            <ListGroup.Item>Injury: {decision.actionData[0]}</ListGroup.Item>
-            <ListGroup.Item>Treatment: {decision.actionData[2]}</ListGroup.Item>
-          </div>
-        )
-      case "TAG_APPLIED":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {decision.actionData[0]}</ListGroup.Item>
-            <ListGroup.Item>Tag: {this.tagMappings[decision.actionData[1]]}</ListGroup.Item>
-          </div>
-        )
-      case "MOVE_TO_EVAC":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {decision.actionData[0]}</ListGroup.Item>
-          </div>
-        )
-      default:
-        return (<p>unrecognized action</p>)
-    }
-  }
-
-  renderDecisionADM = (decision) => {
-    switch (decision.command) {
-      case "Tag Casualty":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {nameMappings[decision.parameters["Casualty ID"]]}</ListGroup.Item>
-            <ListGroup.Item>Tag: {this.formattedActionType(decision.parameters["Tag"])}</ListGroup.Item>
-          </div>
-        )
-      case "Check All Vitals":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {nameMappings[decision.parameters["Casualty ID"]]}</ListGroup.Item>
-            {/* omitted data for now, adm gets all vital information not just pulse
-            <ListGroup.Item>Breathing: {this.formattedActionType(decision.response["breathing"])}</ListGroup.Item>
-            <ListGroup.Item>Conscious: {decision.response["conscious"] ? `Yes` : `No`}</ListGroup.Item>
-            */}
-            <ListGroup.Item>Pulse Reading: {this.admPulseMapping(parseInt(decision.response["hrpmin"]))}</ListGroup.Item>
-            {/*<ListGroup.Item>Mental Status: {this.formattedActionType(decision.response["mental_status"])}</ListGroup.Item>*/}
-          </div>
-        )
-      case "Apply Treatment":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {nameMappings[decision.parameters["Casualty ID"]]}</ListGroup.Item>
-            <ListGroup.Item>Treatment: {this.formattedActionType(decision.parameters["Parameters"]["treatment"])}</ListGroup.Item>
-            <ListGroup.Item>Location: {this.formattedActionType(decision.parameters["Parameters"]["location"])}</ListGroup.Item>
-          </div>
-        )
-      case "Move to EVAC":
-        return (
-          <div>
-            <ListGroup.Item>Patient: {nameMappings[decision.parameters["Casualty ID"]]}</ListGroup.Item>
-          </div>
-        )
-    }
-  }
-
   humanImageToDecisionMapping(decisions, casualties) {
     // map the proper image to a human decision
     if (!casualties || casualties.length === 0) {
@@ -174,7 +35,6 @@ class DecisionList extends React.Component {
     casualties.forEach(casualty => {
       casualtyMap.set(casualty.name, casualty);
     });
-
 
     const updatedDecisions = decisions.map(decision => {
 
@@ -194,7 +54,6 @@ class DecisionList extends React.Component {
           actionData: updatedActionData
         };
       }
-
 
       return decision;
     });
@@ -228,11 +87,11 @@ class DecisionList extends React.Component {
   componentDidUpdate(prevProps) {
     if (this.props.decisionMaker !== prevProps.decisionMaker) {
       this.setState({ decisions: [] })
-      this.setState({ decisionMaker: this.props.decisionMaker})
+      this.setState({ decisionMaker: this.props.decisionMaker })
       this.fetchData(this.props.decisionMaker);
     }
     if (this.props.selectedScenario != prevProps.selectedScenario) {
-      this.setState({ scenario: this.props.selectedScenario})
+      this.setState({ scenario: this.props.selectedScenario })
     }
   }
 
@@ -272,7 +131,6 @@ class DecisionList extends React.Component {
         });
     }
   }
-
 
   parseCSV(csvData) {
     const dataLines = csvData.split('\n');
@@ -326,8 +184,6 @@ class DecisionList extends React.Component {
       });
   }
 
-
-
   parseDoc() {
     const { htmlFileContent } = this.state;
     const parser = new DOMParser();
@@ -342,9 +198,9 @@ class DecisionList extends React.Component {
     let tableArray = Array.from(tableTags)
     if (this.props.selectedScenario === "st-september-2023-mvp2") {
       // have to filter out different actions from simulator based on scenario
-      tableArray = this.filterOutCasualtiesForSoarTech(getCasualtyArray(tableArray))
-    } else if (this.props.selectedScenario === "adept-september-demo-scenario-1"){
-      tableArray = this.filterOutCasualtiesForBBN(getCasualtyArray(tableArray))
+      tableArray = this.filterOutCasualtiesForSoarTech(utility.getCasualtyArray(tableArray))
+    } else if (this.props.selectedScenario === "adept-september-demo-scenario-1") {
+      tableArray = this.filterOutCasualtiesForBBN(utility.getCasualtyArray(tableArray))
     }
 
     this.setState({ casualtyArray: tableArray })
@@ -355,7 +211,7 @@ class DecisionList extends React.Component {
     let decisions = this.humanImageToDecisionMapping(this.state.csvFileContent, this.state.casualtyArray);
     if (this.props.selectedScenario === "st-september-2023-mvp2") {
       decisions = this.filterDecisionsSoarTech(decisions)
-    } else if (this.props.selectedScenario === "adept-september-demo-scenario-1"){
+    } else if (this.props.selectedScenario === "adept-september-demo-scenario-1") {
       decisions = this.filterDecisionsBBN(decisions)
     }
     this.setState({ decisions: decisions })
@@ -424,10 +280,10 @@ class DecisionList extends React.Component {
   }
 
   filterDecisionsADM(history) {
-    const valuesToFilter = 
-    ["Start Session", "Start Scenario", "Get Scenario State", "Take Action", 
-      "Respond to TA1 Probe", "Request SITREP", "TA1 Alignment Target Session ID", "TA1 Alignment Target Data", 
-      "TA1 Probe Response Alignment", "TA1 Session Alignment"];
+    const valuesToFilter =
+      ["Start Session", "Start Scenario", "Get Scenario State", "Take Action",
+        "Respond to TA1 Probe", "Request SITREP", "TA1 Alignment Target Session ID", "TA1 Alignment Target Data",
+        "TA1 Probe Response Alignment", "TA1 Session Alignment"];
 
     const filteredHistory = history.filter(entry => !valuesToFilter.includes(entry.command));
 
@@ -449,24 +305,9 @@ class DecisionList extends React.Component {
           <Accordion style={{ height: accordionHeight, overflowY: 'scroll' }}>
             {this.state.decisions.map((decision, index) => (
               <Accordion.Item key={index} eventKey={index}>
-                <Accordion.Header>{this.humanActionMap[this.formattedActionType(decision.actionType)]}</Accordion.Header>
+                <Accordion.Header>{utility.humanActionMap[utility.formattedActionType(decision.actionType)]}</Accordion.Header>
                 <Accordion.Body>
-                  <div className="row">
-                    <div className="col">
-                      <ListGroup variant="flush">
-                        {this.renderDecisionCSV(decision)}
-                      </ListGroup>
-                    </div>
-                    {decision.imgURL && (
-                      <div className="col">
-                        <img
-                          src={decision.imgURL}
-                          alt="casualty"
-                          className="img-fluid"
-                        />
-                      </div>
-                    )}
-                  </div>
+                  <Decision isHuman={true} decision={decision}></Decision>
                 </Accordion.Body>
               </Accordion.Item>
             ))}
@@ -474,7 +315,7 @@ class DecisionList extends React.Component {
         )}
 
         {this.props.selectedScenario && this.props.decisionMaker && this.state.decisionMaker && !this.props.isHuman && (
-          <Query query={test_by_adm_and_scenario} variables={{"scenarioID": this.props.selectedScenario, "admName": this.props.decisionMaker}}>
+          <Query query={test_by_adm_and_scenario} variables={{ "scenarioID": this.props.selectedScenario, "admName": this.props.decisionMaker }}>
             {({ loading, error, data }) => {
               if (loading) return <div>Loading ...</div>;
               if (error) return <div>Error</div>;
@@ -482,34 +323,19 @@ class DecisionList extends React.Component {
               decisions = this.admImageToDecisionMapping(this.filterDecisionsADM(data.getTestByADMandScenario.history))
               return (
                 <div>
-                {this.state.decisions && (
-                <Accordion style={{ height: accordionHeight, overflowY: 'scroll' }}>
-                  {decisions.map((decision, index) => (
-                    <Accordion.Item key={index} eventKey={index}>
-                      <Accordion.Header>{this.admCommandMap[decision.command]}</Accordion.Header>
-                      <Accordion.Body>
-                        <div className="row">
-                          <div className="col">
-                            <ListGroup variant="flush">
-                              {this.renderDecisionADM(decision)}
-                            </ListGroup>
-                          </div>
-                          {decision.imgURL && (
-                            <div className="col">
-                              <img
-                                src={decision.imgURL}
-                                alt="casualty"
-                                className="img-fluid"
-                              />
-                            </div>
-                          )}
-                        </div>
-                      </Accordion.Body>
-                    </Accordion.Item>
-                  ))}
-                </Accordion>
-              )}
-              </div>
+                  {decisions && (
+                    <Accordion style={{ height: accordionHeight, overflowY: 'scroll' }}>
+                      {decisions.map((decision, index) => (
+                        <Accordion.Item key={index} eventKey={index}>
+                          <Accordion.Header>{utility.admCommandMap[decision.command]}</Accordion.Header>
+                          <Accordion.Body>
+                            <Decision isHuman={false} decision={decision}></Decision>
+                          </Accordion.Body>
+                        </Accordion.Item>
+                      ))}
+                    </Accordion>
+                  )}
+                </div>
               );
             }}
           </Query>
@@ -518,14 +344,6 @@ class DecisionList extends React.Component {
     );
   };
 
-}
-
-export const nameMappings = {
-  "MarineC": "Military Mike Jungle Combat_3_2 Root",
-  "MarineA": "Military Mike Jungle Burned_0 Root",
-  "Intelligence Officer": "Intelligence Officer Burned_Gary_1 Root",
-  "Mike": "Military Mike Jungle Scout_1_3 Root",
-  "Civilian": "Civilian Bob_4 Root"
 }
 
 export default DecisionList;
